lab4/requirements.txt
autopep8==1.5.1
cycler==0.10.0
kiwisolver==1.2.0
matplotlib==3.2.1
numpy==1.18.2
opencv-python==4.2.0.34
Pillow==7.1.1
pkg-resources==0.0.0
pycodestyle==2.5.0
pyparsing==2.4.7
python-dateutil==2.8.1
scipy==1.4.1
six==1.14.0
